# docs/api/service-openapi.yaml
openapi: 3.1.0
info:
  title: TrioBase Governed OpenAPI Integration Service
  version: 0.1.0
  license:
    name: Proprietary
servers:
  - url: /api/v1/openapi
...
security:
  - applicationCredential: []
paths:
...
  /management/products/{id}:
    put:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: updateApiProduct
      summary: Update API product metadata
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: API product }
  /management/products/{id}/versions:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: createApiProductVersion
      summary: Create an API product draft version
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: API product version }
  /management/products/versions/{id}:
    get:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: getApiProductVersion
      summary: Get an API product version
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: API product version }
    put:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: updateApiProductVersion
      summary: Update an API product draft version
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: API product version }
  /management/products/versions/{id}/publish:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: publishApiProductVersion
      summary: Publish an API product version
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: API product version }
  /management/products/versions/{id}/deprecate:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: deprecateApiProductVersion
      summary: Deprecate an API product version
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: API product version }
  /management/products/versions/{id}/archive:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: archiveApiProductVersion
      summary: Archive an API product version
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: API product version }
  /management/products/{id}/archive:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: archiveApiProduct
      summary: Archive an API product
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: API product archived }
  /management/products/{id}/grants/applications/{applicationId}:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: grantApiProductToApplication
      summary: Grant a private API product to an application
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200': { description: Access grant }
  /management/products/discover:
    get:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: discoverApiProducts
      summary: Discover API products available to an application
      parameters:
        - $ref: '#/components/parameters/ApplicationIdQuery'
      responses:
        '200': { description: Discoverable products }
...
  /management/applications/{id}:
    get:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: getIntegrationApplication
      summary: Get an integration application
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application }
    put:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: updateIntegrationApplication
      summary: Update an integration application
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application }
  /management/applications/{id}/submit:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: submitIntegrationApplication
      summary: Submit an application for approval
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application }
  /management/applications/{id}/activate:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: activateIntegrationApplication
      summary: Activate an approved application
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application }
  /management/applications/{id}/suspend:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: suspendIntegrationApplication
      summary: Suspend an application
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application }
  /management/applications/{id}/reactivate:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: reactivateIntegrationApplication
      summary: Reactivate a suspended application
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application }
  /management/applications/{id}/expire:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: expireIntegrationApplication
      summary: Expire an application
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application }
  /management/applications/{id}/revoke:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: revokeIntegrationApplication
      summary: Revoke an application
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application }
  /management/applications/{id}/clients:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: createApplicationClient
      summary: Create an environment-scoped application client
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application client }
  /management/applications/clients/{id}/activate:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: activateApplicationClient
      summary: Activate an application client
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application client }
  /management/applications/clients/{id}/suspend:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: suspendApplicationClient
      summary: Suspend an application client
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application client }
  /management/applications/clients/{id}/reactivate:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: reactivateApplicationClient
      summary: Reactivate an application client
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application client }
  /management/applications/clients/{id}/revoke:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: revokeApplicationClient
      summary: Revoke an application client
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application client }
  /management/applications/clients/{id}/credentials:
    get:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: listApplicationCredentials
      summary: List credential bindings for an application client
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Credential bindings }
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: bindApplicationCredential
      summary: Bind an imported secret reference to an application client
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Credential binding }
  /management/applications/clients/{id}/credentials/rotate:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: rotateApplicationCredential
      summary: Rotate a generated credential and return it once
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Credential binding with one-time material }
  /management/applications/credentials/{id}/revoke:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: revokeApplicationCredential
      summary: Revoke a credential binding and publish fail-closed policy state
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Credential binding }
  /management/applications/credentials/{id}/plaintext:
    get:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: rejectStoredCredentialPlaintextRead
      summary: Always rejects stored credential plaintext retrieval
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '403': { description: Stored credential plaintext is never returned }
...
  /management/subscriptions/{id}/activate:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: activateProductSubscription
      summary: Approve and activate a product subscription
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Subscription }
  /management/subscriptions/{id}/upgrade:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: upgradeProductSubscription
      summary: Request a subscription upgrade to another product version
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Subscription }
  /management/subscriptions/{id}/suspend:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: suspendProductSubscription
      summary: Suspend a product subscription
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Subscription }
  /management/subscriptions/{id}/revoke:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: revokeProductSubscription
      summary: Revoke a product subscription
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Subscription }
  /management/subscriptions/{id}/expire:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: expireProductSubscription
      summary: Expire a product subscription
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Subscription }
  /management/approvals/{id}/decision:
    post:
      tags: [Applications]
      x-triobase-service: service-openapi
      operationId: decideAssetApproval
      summary: Approve or reject an application or product access request
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Approval decision }
...
  /management/policies/{id}:
    put:
      tags: [Policies]
      x-triobase-service: service-openapi
      operationId: updateTrafficPolicy
      summary: Update a draft traffic policy
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Traffic policy version }
  /management/policies/{id}/publish:
    post:
      tags: [Policies]
      x-triobase-service: service-openapi
      operationId: publishTrafficPolicy
      summary: Publish a traffic policy version
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Traffic policy version }
  /management/policies/snapshots:
    post:
      tags: [Policies]
      x-triobase-service: service-openapi
      operationId: publishPolicySnapshot
      summary: Publish a signed immutable policy snapshot
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/EnvironmentQuery'
      responses:
        '200': { description: Policy snapshot }
  /management/policies/enforcement/{point}/applied:
    post:
      tags: [Policies]
      x-triobase-service: service-openapi
      operationId: reportPolicySnapshotApplied
      summary: Report that an enforcement point applied a policy snapshot
      parameters:
        - $ref: '#/components/parameters/Point'
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/EnvironmentQuery'
        - name: version
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        '200': { description: Enforcement state }
  /management/policies/enforcement/{point}:
    get:
      tags: [Policies]
      x-triobase-service: service-openapi
      operationId: getPolicyEnforcementState
      summary: Get policy snapshot application state for an enforcement point
      parameters:
        - $ref: '#/components/parameters/Point'
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/EnvironmentQuery'
      responses:
        '200': { description: Enforcement state }
...
components:
  securitySchemes:
    applicationCredential:
      type: apiKey
      in: header
      name: X-Client-Credential
      description: Gateway-authenticated application credential plus active subscription and policy.
    callbackSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC or RSA signature validated with timestamp and nonce replay protection.
    gatewayContext:
      type: apiKey
      in: header
      name: X-OpenAPI-Gateway-Secret
      description: Shared gateway/runtime trust secret. Production callers must also send service identity headers.
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string }
...
    Point:
      name: point
      in: path
      required: true
      schema: { type: string }
...
    TenantIdQuery:
      name: tenantId
      in: query
      required: false
      schema: { type: string }
    EnvironmentQuery:
      name: environment
      in: query
      required: true
      schema:
        type: string
        enum: [DEV, TEST, PROD]
    ApplicationIdQuery:
      name: applicationId
      in: query
      required: true
      schema: { type: string }
...